Show CSV export dialog again when permissions are granted. Closes #398.
